rewriting main, refactored environmental data transformation and pond sizing
